move scipio file to its own file
Plus move common operations to it. The goal is implementing buffered files
soon
